Change in QtDesigner classes of plot widgets.
Now there is Base classes for all QtDesigner classes.
