+++ pydm/widgets/timeplot_curve_editor.py
@@ -1,87 +1,52 @@
-from ..PyQt.QtGui import QDialog, QVBoxLayout, QHBoxLayout, QTableView, QAbstractItemView, QSpacerItem, QSizePolicy, QDialogButtonBox, QPushButton, QItemSelection, QStyledItemDelegate, QColorDialog
-from ..PyQt.QtCore import Qt, pyqtSlot, QModelIndex
-from ..PyQt.QtDesigner import QDesignerFormWindowInterface
-from .timeplot_table_model import PyDMTimePlotCurvesModel
+from ..PyQt.QtCore import QModelIndex, QVariant
+from .baseplot_table_model import BasePlotCurvesModel
+from .baseplot_curve_editor import BasePlotCurveEditorDialog
 
-class TimePlotCurveEditorDialog(QDialog):
+
+class PyDMTimePlotCurvesModel(BasePlotCurvesModel):
+
     def __init__(self, plot, parent=None):
-        super(TimePlotCurveEditorDialog, self).__init__(parent)
-        self.plot = plot
-        self.setup_ui()
-        self.column_names = ("Channel", "Label", "Color")
-        self.table_model = PyDMTimePlotCurvesModel(self.plot)
-        self.table_view.setModel(self.table_model)
-        self.table_model.plot = plot
-        #self.table_view.resizeColumnsToContents()
-        self.add_button.clicked.connect(self.addCurve)
-        self.remove_button.clicked.connect(self.removeSelectedCurve)
-        self.remove_button.setEnabled(False)
-        self.table_view.selectionModel().selectionChanged.connect(self.handleSelectionChange)
-        
-    def setup_ui(self):
-        self.resize(500, 300)
-        self.vertical_layout = QVBoxLayout(self)
-        self.table_view = QTableView(self)
-        self.table_view.setEditTriggers(QAbstractItemView.DoubleClicked)
-        self.table_view.setProperty("showDropIndicator", False)
-        self.table_view.setDragDropOverwriteMode(False)
-        self.table_view.setSelectionMode(QAbstractItemView.SingleSelection)
-        self.table_view.setSelectionBehavior(QAbstractItemView.SelectRows)
-        color_delegate = ColorColumnDelegate(self)
-        self.table_view.setItemDelegateForColumn(2, color_delegate)
-        self.table_view.doubleClicked.connect(self.handleDoubleClick)
-        self.table_view.setSortingEnabled(False)
-        self.table_view.horizontalHeader().setStretchLastSection(True)
-        self.table_view.verticalHeader().setVisible(False)
-        self.table_view.setColumnWidth(0, 166)
-        self.table_view.setColumnWidth(1, 166)
-        self.table_view.setColumnWidth(2, 166)
-        self.vertical_layout.addWidget(self.table_view)
-        self.add_remove_layout = QHBoxLayout()
-        spacer = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
-        self.add_remove_layout.addItem(spacer)
-        self.add_button = QPushButton("Add Curve", self)
-        self.add_remove_layout.addWidget(self.add_button)
-        self.remove_button = QPushButton("Remove Curve", self)
-        self.add_remove_layout.addWidget(self.remove_button)
-        self.vertical_layout.addLayout(self.add_remove_layout)
-        self.button_box = QDialogButtonBox(self)
-        self.button_box.setOrientation(Qt.Horizontal)
-        self.button_box.addButton("Done", QDialogButtonBox.AcceptRole)
-        self.vertical_layout.addWidget(self.button_box)
-        self.button_box.accepted.connect(self.saveChanges)
-        self.button_box.rejected.connect(self.reject)
-        self.setWindowTitle("Timeplot Curve Editor")
-    
-    @pyqtSlot()
-    def addCurve(self):
-        self.table_model.append("")
-    
-    @pyqtSlot()
-    def removeSelectedCurve(self):
-        self.table_model.removeAtIndex(self.table_view.currentIndex())
-        
-    @pyqtSlot(QItemSelection, QItemSelection)
-    def handleSelectionChange(self, selected, deselected):
-        self.remove_button.setEnabled(self.table_view.selectionModel().hasSelection())
-    
-    @pyqtSlot(QModelIndex)
-    def handleDoubleClick(self, index):
-        if self.table_model.needsColorDialog(index):
-            #The table model returns a QBrush for BackgroundRole, not a QColor.
-            init_color = self.table_model.data(index, Qt.BackgroundRole).color()
-            color = QColorDialog.getColor(init_color, self)
-            if color.isValid():
-                self.table_model.setData(index, color, role=Qt.EditRole)
-    
-    @pyqtSlot()
-    def saveChanges(self):
-        formWindow = QDesignerFormWindowInterface.findFormWindow(self.plot)
-        if formWindow:
-            formWindow.cursor().setProperty("curves", self.plot.curves)
-        self.accept()
-    
-class ColorColumnDelegate(QStyledItemDelegate):
-    def createEditor(self, parent, option, index):
-        return None
-        
+        super(PyDMTimePlotCurvesModel, self).__init__(plot, parent=parent)
+        self._column_names = ("Channel", ) + self._column_names
+
+    def get_data(self, column_name, curve):
+        if column_name == "Channel":
+            if curve.address is None:
+                return QVariant()
+            return str(curve.address)
+        return super(PyDMTimePlotCurvesModel, self).get_data(column_name,
+                                                             curve)
+
+    def set_data(self, column_name, curve, value):
+        if column_name == "Channel":
+            curve.address = str(value)
+        else:
+            return super(PyDMTimePlotCurvesModel, self).set_data(
+                            column_name=column_name, curve=curve, value=value)
+        return True
+
+    def append(self, address=None, name=None, color=None):
+        self.beginInsertRows(QModelIndex(), len(self._plot._curves),
+                             len(self._plot._curves))
+        self._plot.addYChannel(address, name, color)
+        self.endInsertRows()
+
+    def removeAtIndex(self, index):
+        self.beginRemoveRows(QModelIndex(), index.row(), index.row())
+        self._plot.removeYChannelAtIndex(index.row())
+        self.endRemoveRows()
+
+
+class TimePlotCurveEditorDialog(BasePlotCurveEditorDialog):
+    """TimePlotCurveEditorDialog is a QDialog that is used in Qt Designer
+    to edit the properties of the curves in a waveform plot.  This dialog is
+    shown when you double-click the plot, or when you right click it and
+    choose 'edit curves'.
+
+    This thing is mostly just a wrapper for a table view, with a couple
+    buttons to add and remove curves, and a button to save the changes."""
+    TABLE_MODEL_CLASS = PyDMTimePlotCurvesModel
+
+    def __init__(self, plot, parent=None):
+        super(TimePlotCurveEditorDialog, self).__init__(plot, parent)
+        self.setup_delegate_columns(index=2)
